+++ tests/test_components.py
@@ -53,17 +53,17 @@ except ModuleNotFoundError:
 
 
 class TestBaseContext:
-    def test_id_match_property(self):
-        context = yuyo.components.Context(
-            mock.Mock(), mock.Mock(), "yeyeye meow meow", "", register_task=lambda v: None
-        )
+    def test_author_property(self):
+        mock_interaction = mock.Mock()
+        context = yuyo.components.Context( mock.Mock(), mock_interaction, "", "", register_task=lambda v: None)
 
-        assert context.id_match == "yeyeye meow meow"
+        assert context.author is mock_interaction.user
 
-    def test_id_metadata_property(self):
-        context = yuyo.components.Context(mock.Mock(), mock.Mock(), "", "very meta girl", register_task=lambda v: None)
+    def test_created_at_property(self):
+        mock_interaction = mock.Mock()
+        context = yuyo.components.Context( mock.Mock(), mock_interaction, "", "", register_task=lambda v: None)
 
-        assert context.id_metadata == "very meta girl"
+        assert context.created_at is mock_interaction.created_at
 
     @pytest.mark.parametrize(
         ("now", "expires_at"),
@@ -88,6 +88,24 @@ class TestBaseContext:
 
             assert context.expires_at == expires_at
 
+    def test_guild_id_property(self):
+        mock_interaction = mock.Mock()
+        context = yuyo.components.Context( mock.Mock(), mock_interaction, "", "", register_task=lambda v: None)
+
+        assert context.guild_id is mock_interaction.guild_id
+
+    def test_id_match_property(self):
+        context = yuyo.components.Context(
+            mock.Mock(), mock.Mock(), "yeyeye meow meow", "", register_task=lambda v: None
+        )
+
+        assert context.id_match == "yeyeye meow meow"
+
+    def test_id_metadata_property(self):
+        context = yuyo.components.Context(mock.Mock(), mock.Mock(), "", "very meta girl", register_task=lambda v: None)
+
+        assert context.id_metadata == "very meta girl"
+
     def test_interaction_property(self):
         mock_interaction = mock.Mock()
 
@@ -96,6 +114,11 @@ class TestBaseContext:
         )
 
         assert context.interaction is mock_interaction
+    def test_member_property(self):
+        mock_interaction = mock.Mock()
+        context = yuyo.components.Context( mock.Mock(), mock_interaction, "", "", register_task=lambda v: None)
+
+        assert context.member is mock_interaction.member
 
 
 class TestComponentContext:
